🚚 Rename: 타입명 정리
